do not reference IElement in ChangeAttributeEdit
When an element is deleted, it gets dissociated from the document. While
it can still be referenced by an ChangeAttributeEdit, it is useless for
undo/redo, because its ContentRange gets lost when it is dissociated.
Hence instead of referencing the element, the document and the current
caret offset where the edit took place are stored in
ChangeAttributeEdit.

Signed-off-by: Florian Thienel <florian@thienel.org>