StyleSheet caching changed for better performance.
The former caching implementation used WeakReference for cached styles.
As styles are not referenced anywhere else, the cache was actually
useless because the garbage collector removed the styles every few
seconds.
This change also fixes the problem with E + F CSS selectors, that were
not applied while editing a document.

Change-Id: If6e56cebd4fb4e42390b158e2d86585099f3dfa3
Signed-off-by: Carsten Hiesserich <carsten.hie@gmail.com>