bug 418091: synchronize document when Vex gets activated or deactivated
With this change, Vex should integrate better with the Eclipse
ecosystem. VexEditor does not operate on the input directly, but uses a
document provider as an intermediate (as all other Eclipse editors also
do). The integration is not perfect. Vex still uses it's internal
GapContent that is not synchronized with the document provider. This
results in two problems:
    The caret position gets lost when switching editors
    The undo history gets lost when switching editors

Change-Id: Idac0ff8022a6477fc99594f482cfd28f37f86e85
Signed-off-by: Carsten Hiesserich <carsten.hie@gmail.com>