move decision isBlock/isInline solely to CssWhitespacePolicy
There was a duplication in Stlyes.isBlock() and
AbstractBlockBox.isInline(node). Both methods decided if a node was
block or inline, but they were not consistent. Since the isBlock
property depends also on the display property of the parent node, the
responsibility was moved to the CssWhitespacePolicy. The information in
Styles was not sufficient.

Change-Id: Ic6a9f590009a1749433db0494e14bb7b5be78c89