[423130] use an IContentDescriber to associate Vex with files
Vex should be associated only with files containing documents of
registered document types. Therefor the VexContentDescriber reads the
DTD and the root element to decide if a file contains such a document.

Furthermore it is not essentially required for bundles that register a
document type to also register a new content type. This is only
necessary if a custom file extension (!= xml) should be used.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=423130
Change-Id: I695e512c03f49f87f6221d3c3c6529ea04935477
Signed-off-by: Florian Thienel <florian@thienel.org>