make IInlineBox coordinates relative to their containing Paragraph
For a IInlineBox, the Paragraph instance is the parent box in the box
structure. The LineArrangement and Line are only utilities that are used
internally by Paragraph. To be consistent with rest of the box
structure, the coordinates for top and left of an IInlineBox should be
relative to its parent, i.e. the Paragraph.

Signed-off-by: Florian Thienel <florian@thienel.org>