fix error message and missing doctypes in .css file property page
The property page expected all document types to have a public id. As
this is not the case for XML-Schemas (Vex uses the namespace name here),
a NullPointerException was raised that caused the property dialog to
show an error 'The currently displayed page contains illegal values'.
Also all document types after the first XML-Schema were not displayed.

With this change, the attribute 'publicId' in the <doctypeRef>
element of the project xml is renamed to 'doctypeId'. For backward
compatibility, the 'publicId' is still recognized when reading the
config.

Change-Id: Iaa9869c3b81bc9347b62a417e214b6fdc705954f
Signed-off-by: Carsten Hiesserich <carsten.hie@gmail.com>