add URI attribute to plugin doctype definition
The build-in document types use catalog contributions to map public or
system id's to the dtd file. In plugin projects, the user defined
document types had to use the file name as system id to be resolved.
This commit adds an additional 'uri' attribute to the plugin config to
config the file name independently.

Change-Id: Ib26e3b93607a2a4d54dcb3b2ae146f8d1e8db2d7
Signed-off-by: Carsten Hiesserich <carsten.hie@gmail.com>