include PDE into target platform to work around the eternal launch times
Launching the Vex plug-in tests in the IDE requires the plug-in
org.eclipse.pde.junit.runtime. If the JUnitLaunchConfigurationDelegate
cannot find this plug-in in the target platform, it searches in the p2
cache. This can take very long (> 10 minutes) if you have the plug-ins
of several installations in the p2 cache.

To avoid this long launch times, we add the PDE to the target platform.

Change-Id: Ib5dbceccc0c2ed151b2dfa32897785a3eff4f4c6
Signed-off-by: Florian Thienel <florian@thienel.org>